% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_LN.R
\docType{data}
\name{LN_throat_info}
\alias{LN_throat_info}
\title{The sample template data for simulation - Logit-normal Multinomial}
\format{
List of 4.
}
\usage{
data(LN_throat_info)
}
\description{
The data before it's simulated with LN model in three scenarios:
common lineage, rare lineage, random OTUs.
}
\examples{
data(LN_throat_info)

}
\keyword{datasets}
