% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS.R
\name{MDS}
\alias{MDS}
\title{Cluster OTU Data and Evaluate the Result with Adjusted ARI}
\usage{
MDS(otu.count, distance, tree, mdsrank, k, est, threshold = 0.9,
           K.max, group, randMethod = "MA", seed = 2021)
}
\arguments{
\item{otu.count}{OTUs with count numbers or a data matrix. The samples are the rows.}

\item{distance}{character string specifying the distance to be used for calculating dissimilarities between observations.
The currently available options are "bray", "jaccard", "unweighted UniFrac",
"weighted UniFrac", "generalized UniFrac", "PhILR" and "clr" which are corresponding to Bray-Curtis distance, Jaccard Index,
unweighted, weighted, generalized UniFrac, Euclidean after Phylogenetic ILR (PhILR) transformation and center logratio (CLR) transformation.}

\item{tree}{specifies the corresponding tree if choose "unweighted UniFrac", "weighted UniFrac", "generalized UniFrac" or "PhILR" as the distance.}

\item{mdsrank}{the maximum dimension of the space which the data are to be represented in when conduct the MDS; must be less than or equal to n-1.}

\item{k}{true number of the clusters. If not specified, please choose one of the three methods using parameter "est" to estimate the number.}

\item{est}{specifies the preferred methods to estimate the number of clusters. Available choices are "gap" (Gap statistic), "ps" (Prediction Strength), "si" (Silhouette Index).}

\item{threshold}{the threshold for prediction strength method mentioned in "est" taking value from 0 to 1. By default, it's 0.9. In practice, 0.8 or a higher one is preferred.}

\item{K.max}{integer. Specify the maximum number of clusters if use some methods to estimate k.}

\item{group}{a numeric vector of class labels. The length of group should be the same as the number of samples.}

\item{randMethod}{specifies the preferred external index to measure the agreement between the clustering result and the given group.
Available indices are:  "Rand", "HA" (Hubert and Arabie's adjusted Rand index), "MA" (Morey and Agresti's adjusted Rand index), "FM" (Fowlkes and Mallows's index), "Jaccard" (Jaccard index).
By default, "MA" will be output.}

\item{seed}{set.seed with the given number. By default, set.seed(2021).}
}
\value{
A list of 3 containing bestk (the best number of clusters), ARI (the corresponding adjusted
Rand Index) and cluster (the clustering results).
}
\description{
This function can conduct classical multidimensional scaling (MDS) with
seven different distances and then calculate agreement indices between any
two partitions for a data set.
}
\examples{
data(Martinez)
tree = Martinez$tree
otu.table = Martinez$otutable
otu.count = t(otu.table)
MDS(otu.count, distance = "bray", mdsrank = 5, est = "gap", K.max = 16,
group = as.numeric(as.factor(Martinez$sampleinfo)), randMethod = "MA")
}
\keyword{ARI}
\keyword{MDS,}
