% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Smits.R
\docType{data}
\name{Smits}
\alias{Smits}
\title{Smits Data}
\format{
A large list with 3 components which are tree, sampleinfo and otutable.
}
\usage{
data(Smits)
}
\description{
Smits data is a data set from the R package "MicrobiomeCluster" including OTU table,
tree and sample seasonal information. The samples are from the early wet season and the late dry season.
}
\examples{
data(Smits)
table(Smits$sampleinfo)

}
\references{
Smits, S. A., Leach, J., Sonnenburg, E. D., Gonzalez, C. G., Lichtman, J. S., Reid, G., Knight, R., Manjurano, A., Changalucha, J., Elias, J. E., Dominguez-Bello, M. G., and Sonnenburg, J. L. (2017). Seasonal cycling in the gut microbiome of the hadza hunter- gatherers of Tanzania. Science, 357(6353):802-806.
}
\keyword{datasets}
